DDF-2507 removed unnecessary attribute registration during ingestion
