+++ catalog/transformer/catalog-transformer-streaming-impl/src/test/java/org/codice/ddf/transformer/xml/streaming/impl/TestXmlInputTransformer.java
@@ -21,7 +21,6 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyInt;
 import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -38,10 +37,8 @@ import org.codice.ddf.parser.xml.XmlParser;
 import org.codice.ddf.transformer.xml.streaming.Gml3ToWkt;
 import org.codice.ddf.transformer.xml.streaming.SaxEventHandler;
 import org.codice.ddf.transformer.xml.streaming.SaxEventHandlerFactory;
-import org.codice.ddf.transformer.xml.streaming.lib.MetacardTypeRegister;
 import org.codice.ddf.transformer.xml.streaming.lib.SaxEventHandlerDelegate;
 import org.codice.ddf.transformer.xml.streaming.lib.XmlInputTransformer;
-import org.junit.Before;
 import org.junit.Test;
 import org.xml.sax.Attributes;
 import org.xml.sax.ErrorHandler;
@@ -71,14 +68,6 @@ public class TestXmlInputTransformer {
     static InputStream inputStream;
 
     XmlInputTransformer xmlInputTransformer;
-
-    MetacardTypeRegister mockMetacardTypeRegister;
-
-    @Before
-    public void setup() {
-        mockMetacardTypeRegister = mock(MetacardTypeRegister.class);
-        doReturn(BasicTypes.BASIC_METACARD).when(mockMetacardTypeRegister).getMetacardType();
-    }
 
     /*
         Tests a base XmlInputTransformer, CONTENT_TYPE is null because it is not in the base xmlToMetacard mapping
@@ -216,7 +205,6 @@ public class TestXmlInputTransformer {
         inputStream = new FileInputStream("src/test/resources/metacard1.xml");
         xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("gml-handler"));
-        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
         GmlHandlerFactory factory = new GmlHandlerFactory();
         factory.setGml3ToWkt(gml3ToWkt);
         xmlInputTransformer.setSaxEventHandlerFactories(Collections.singletonList((SaxEventHandlerFactory) factory));
@@ -233,7 +221,6 @@ public class TestXmlInputTransformer {
         inputStream = new FileInputStream("src/test/resources/metacard1.xml");
         xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("gml-handler"));
-        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
         GmlHandlerFactory factory = new GmlHandlerFactory();
         Gml3ToWkt badGml3toWkt = mock(Gml3ToWkt.class);
         when(badGml3toWkt.convert(anyString())).thenThrow(new ValidationExceptionImpl());
