+++ catalog/transformer/catalog-transformer-streaming-lib/src/main/java/org/codice/ddf/transformer/xml/streaming/lib/XmlInputTransformer.java
@@ -19,7 +19,10 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.codice.ddf.transformer.xml.streaming.SaxEventHandler;
@@ -27,15 +30,18 @@ import org.codice.ddf.transformer.xml.streaming.SaxEventHandlerFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import ddf.catalog.data.AttributeDescriptor;
 import ddf.catalog.data.Metacard;
+import ddf.catalog.data.MetacardType;
 import ddf.catalog.data.impl.AttributeImpl;
+import ddf.catalog.data.impl.BasicTypes;
 import ddf.catalog.transform.CatalogTransformerException;
 import ddf.catalog.transform.InputTransformer;
 import ddf.catalog.util.Describable;
 
 /**
  * A {@link InputTransformer} that can be configured to parse any XML into a {@link Metacard}
- * It is configured through {@link XmlInputTransformer#setSaxEventHandlerConfiguration(List)}}
+ * It is configured through {@link XmlInputTransformer#setSaxEventHandlerConfiguration(List)} and {@link XmlInputTransformer#getMetacardType()}
  * {@inheritDoc}
  */
 public class XmlInputTransformer implements InputTransformer, Describable {
@@ -68,11 +74,9 @@ public class XmlInputTransformer implements InputTransformer, Describable {
      */
     private List<String> saxEventHandlerConfiguration;
 
-    private MetacardTypeRegister dynamicMetacardTypeRegister;
-
     /**
      * Method to create a new {@link SaxEventHandlerDelegate}, configured to parse a metacard
-     * according to {@link XmlInputTransformer#saxEventHandlerConfiguration}.
+     * according to {@link XmlInputTransformer#saxEventHandlerConfiguration} and {@link XmlInputTransformer#getMetacardType}
      *
      * @return a new SaxEventHandlerDelegate
      */
@@ -89,8 +93,7 @@ public class XmlInputTransformer implements InputTransformer, Describable {
          * Pass all the new handlers to configure and create a new SaxEventHandlerDelegate and sets
          * the metacardType
          */
-        return new SaxEventHandlerDelegate(filteredSaxEventHandlers).setMetacardType(
-                dynamicMetacardTypeRegister.getMetacardType());
+        return new SaxEventHandlerDelegate(filteredSaxEventHandlers).setMetacardType(getMetacardType());
 
     }
 
@@ -192,14 +195,6 @@ public class XmlInputTransformer implements InputTransformer, Describable {
         this.saxEventHandlerConfiguration = saxEventHandlerConfiguration;
     }
 
-    public void setDynamicMetacardTypeRegister(MetacardTypeRegister metacardTypeRegister) {
-        this.dynamicMetacardTypeRegister = metacardTypeRegister;
-    }
-
-    public MetacardTypeRegister getDynamicMetacardTypeRegister() {
-        return dynamicMetacardTypeRegister;
-    }
-
     @Override
     public String getVersion() {
         return version;
@@ -244,4 +239,27 @@ public class XmlInputTransformer implements InputTransformer, Describable {
     public void setOrganization(String organization) {
         this.organization = organization;
     }
+
+    /**
+     * Defines and returns a {@link DynamicMetacardType} based on component Sax Event Handler Factories
+     * and what attributes they populate
+     *
+     * @return a DynamicMetacardType that describes the type of metacard that is created in this transformer
+     */
+
+    public MetacardType getMetacardType() {
+        Set<AttributeDescriptor> attributeDescriptors =
+                new HashSet<>(BasicTypes.BASIC_METACARD.getAttributeDescriptors());
+
+        if (saxEventHandlerConfiguration != null && saxEventHandlerFactories != null) {
+            attributeDescriptors.addAll(saxEventHandlerFactories.stream()
+                    .filter(p -> saxEventHandlerConfiguration.contains(p.getId()))
+                    .map(SaxEventHandlerFactory::getSupportedAttributeDescriptors)
+                    .flatMap(Collection::stream)
+                    .collect(Collectors.toSet()));
+        }
+
+        return new DynamicMetacardType(attributeDescriptors, id);
+    }
+
 }
